requirements.txt
streamlit
pandas
requests
matplotlibsourc